+++ src/devtools/components/WatchedQueries/WatchedQueries.js
@@ -179,13 +179,21 @@ class WatchedQuery extends React.Component {
       query.metadata &&
       query.metadata.reactComponent &&
       query.metadata.reactComponent.displayName;
+    const componentDisplayName =
+      query &&
+      query.metadata &&
+      query.metadata.component &&
+      query.metadata.component.displayName;
+    const displayName = componentDisplayName || reactComponentDisplayName;
+
     const queryString = query.queryString || query.document.loc.source.body;
+
     return (
       <div className={classnames("main", { loading: query.loading })}>
         <div className="panel-title">
           {queryLabel(queryId, query)}
-          {reactComponentDisplayName && (
-            <span className="component-name">{`<${reactComponentDisplayName}>`}</span>
+          {displayName && (
+            <span className="component-name">{`<${displayName}>`}</span>
           )}
           <span
             className="run-in-graphiql-link"